+++ libraries/PCA9632/PCA9632.h
@@ -3,8 +3,8 @@
 //    FILE: PCA9632.h
 //  AUTHOR: Rob Tillaart
 //    DATE: 2024-11-25
-// VERSION: 0.1.0
-// PURPOSE: Arduino library for PCA9632 and PCA9633 4 channel, I2C LED driver.
+// VERSION: 0.1.1
+// PURPOSE: Arduino library for PCA9632 and PCA9633 I2C 8 bit PWM LED driver, 4 channel.
 //     URL: https://github.com/RobTillaart/PCA9632
 
 
@@ -12,7 +12,7 @@
 #include "Wire.h"
 
 
-#define PCA9632_LIB_VERSION         (F("0.1.0"))
+#define PCA9632_LIB_VERSION         (F("0.1.1"))
 
 //  REGISTERS PCA9632
 #define PCA9632_MODE1               0x00
@@ -23,20 +23,13 @@
 #define PCA9632_PWM3                0x05
 #define PCA9632_GRPPWM              0x06
 #define PCA9632_GRPFREQ             0x07
+
 #define PCA9632_LEDOUT              0x08
 #define PCA9632_SUBADR1             0x09
 #define PCA9632_SUBADR2             0x0A
 #define PCA9632_SUBADR3             0x0B
+#define PCA9632_SUBADR(x)           (0x08 +(x))  //  x = 0..3
 #define PCA9632_ALLCALLADR          0x0C
-
-
-//  REGISTER CONTENT
-//  TODO GRPPWM
-//  TODO GRPFREQ
-//  TODO SUBADR1
-//  TODO SUBADR2
-//  TODO SUBADR3
-//  TODO ALLCALLADR
 
 
 //  ERROR CODES - compatible with PCA9634
@@ -68,7 +61,7 @@
 #define PCA9632_MODE2_ACK           0x08  //  0 = on STOP      1 = on ACK
 #define PCA9632_MODE2_TOTEMPOLE     0x04  //  0 = open drain   1 = totem-pole
 #define PCA9632_MODE2_NONE          0x00
-#define PCA9632_MODE2_DEFAULT       0x01
+#define PCA9632_MODE2_DEFAULT       0x02  //  fix #2
 
 //  modi for LEDOUT REGISTER (4x shifted))
 #define PCA9632_LEDOFF              0x00  //  default @ startup
@@ -83,10 +76,37 @@ class PCA9632
 public:
   explicit PCA9632(const uint8_t deviceAddress, TwoWire *wire = &Wire);
 
-  bool     begin(uint8_t mode1_mask = PCA9632_MODE1_DEFAULT, 
+  bool     begin(uint8_t mode1_mask = PCA9632_MODE1_DEFAULT,
                  uint8_t mode2_mask = PCA9632_MODE2_DEFAULT);
   bool     isConnected();
   uint8_t  getAddress();
+
+
+  /////////////////////////////////////////////////////
+  //
+  //  CONFIGURATION
+  //
+  uint8_t  channelCount();
+
+
+  /////////////////////////////////////////////////////
+  //
+  //  WRITE MODE REGISTERS
+  //
+  uint8_t setMode1(uint8_t value);
+  uint8_t setMode2(uint8_t value);
+  uint8_t getMode1();
+  uint8_t getMode2();
+
+
+  /////////////////////////////////////////////////////
+  //
+  //  GROUP REGISTERS
+  //
+  uint8_t setGroupPWM(uint8_t value);
+  uint8_t getGroupPWM();
+  uint8_t setGroupFREQ(uint8_t value);
+  uint8_t getGroupFREQ();
 
 
   /////////////////////////////////////////////////////
@@ -98,21 +118,35 @@ public:
   uint8_t  writeG(uint8_t G);
   uint8_t  writeB(uint8_t B);
   uint8_t  writeW(uint8_t W);
+  //  if not thinking in RGBW but in channels 0..3
   uint8_t  write(uint8_t channel, uint8_t value);
 
   //  RGBW setting, write four PWM registers, last has default to get "writeRGB()"
   uint8_t  write(uint8_t R, uint8_t G, uint8_t B, uint8_t W = 0);
   uint8_t  write(uint8_t * arr);  //  array of at least 4 elements.
+  uint8_t  allOff();
 
 
   /////////////////////////////////////////////////////
   //
-  //  WRITE MODE REGISTERS
+  //  SUB CALL
+  //  not tested yet, synced from PCA9634
   //
-  uint8_t setMode1(uint8_t value);
-  uint8_t setMode2(uint8_t value);
-  uint8_t getMode1();
-  uint8_t getMode2();
+  //  nr = { 1, 2, 3 }
+  bool    enableSubCall(uint8_t nr);
+  bool    disableSubCall(uint8_t nr);
+  bool    isEnabledSubCall(uint8_t nr);
+  bool    setSubCallAddress(uint8_t nr, uint8_t address);
+  uint8_t getSubCallAddress(uint8_t nr);
+  //
+  //  ALL CALL
+  //  not tested yet, synced from PCA9634
+  //
+  bool    enableAllCall();
+  bool    disableAllCall();
+  bool    isEnabledAllCall();
+  bool    setAllCallAddress(uint8_t address);
+  uint8_t getAllCallAddress();
 
 
   /////////////////////////////////////////////////////
@@ -132,20 +166,14 @@ public:
   int      lastError();
 
 
-  /////////////////////////////////////////////////////
-  //
-  //  OTHER
-  //
-  //  get/setRegister() will be protected in the future
-  //  check datasheet.
-  uint8_t  writeReg(uint8_t reg, uint8_t mask);
-  uint8_t  readReg(uint8_t reg);
-
 protected:
+  //  DIRECT CONTROL
+  uint8_t  writeRegister(uint8_t reg, uint8_t value);  //  returns error status.
+  uint8_t  readRegister(uint8_t reg);
+
   uint8_t _address;
   TwoWire * _wire;
-
-  int _error = PCA9632_OK;
+  int     _error = PCA9632_OK;
 };